fix: Replacing floating point values by integers for prices (#33)
* feat: WIP replacing float by ints for prices

* feat: Replaced other occurrences of prices and resources

+++ database/bsgo/migrations/3_create_players.up.sql
@@ -16,7 +16,7 @@ CREATE TABLE player (
 CREATE TABLE player_resource (
   player INTEGER NOT NULL,
   resource INTEGER NOT NULL,
-  amount NUMERIC(12, 2) NOT NULL,
+  amount INTEGER NOT NULL,
   PRIMARY KEY (player, resource),
   FOREIGN KEY (player) REFERENCES player(id),
   FOREIGN KEY (resource) REFERENCES resource(id)
